s/long/int/ for default_value in enum and flags pspecs.
Thu Jun 21 02:43:10 2001  Tim Janik  <timj@gtk.org>

        * gparamspecs.h: s/long/int/ for default_value in enum and flags
        pspecs.

+++ gobject/gparamspecs.h
@@ -155,14 +155,14 @@ struct _GParamSpecEnum
   GParamSpec    parent_instance;
   
   GEnumClass   *enum_class;
-  glong         default_value;
+  gint          default_value;
 };
 struct _GParamSpecFlags
 {
   GParamSpec    parent_instance;
   
   GFlagsClass  *flags_class;
-  gulong        default_value;
+  guint         default_value;
 };
 struct _GParamSpecFloat
 {
